[CHORE]: Organized files
